refactor: much needed re-organization of code
