+++ src/Components/Footer.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { FaInstagram, FaGithub, FaEnvelope } from 'react-icons/fa';
-import { Link } from 'react-router-dom';
 
 const Footer = () => {
   return (
