+++ src/pages/UpcomingProjects.jsx
@@ -7,17 +7,24 @@ const UpcomingProjects = () => {
         WebkitBackgroundClip: 'text',
         WebkitTextFillColor: 'transparent',
         animation: 'bounce 2s infinite',
-        fontFamily: "'Roboto Mono', monospace"
+        fontFamily: "'Roboto Mono', monospace",
     };
 
     return (
         <div className="h-[91.1vh] w-full font-monoS flex flex-col items-center justify-center bg-slate-900">
             <div className="text-center">
-                <h1 style={gradientTextStyle} className="text-7xl font-mono mb-6 font-bold mt-8">Upcoming Projects</h1>
-                <h2 className="text-xl text-gray-300 mt-4 font-mono animate-pulse">Coming Soon</h2>
+                <h1
+                    style={gradientTextStyle}
+                    className="text-7xl font-mono mb-6 font-bold mt-8"
+                >
+                    Upcoming Projects
+                </h1>
+                <h2 className="text-xl text-gray-300 mt-4 font-mono animate-pulse">
+                    Coming Soon
+                </h2>
             </div>
         </div>
     );
-}
+};
 
 export default UpcomingProjects;
